+++ salem/utils.py
@@ -327,6 +327,29 @@ def joblib_read_img_url(url):
     return imread(io.BytesIO(fd.read()))
 
 
+def nice_scale(mapextent, maxlen=0.15):
+    """Returns a nice number for a legend scale of a map.
+
+    Parameters
+    ----------
+    mapextent : float
+        the total extent of the map
+    maxlen : float
+        from 0 to 1, the maximum relative length allowed for the scale
+
+    Examples
+    --------
+    >>> print(nice_scale(140))
+    20.0
+    >>> print(nice_scale(140, maxlen=0.5))
+    50.0
+    """
+    d = np.array([1, 2, 5])
+    e = (np.ones(12) * 10) ** (np.arange(12)-5)
+    candidates = np.matmul(e[:, None],  d[None, :]).flatten()
+    return np.max(candidates[candidates / mapextent <= maxlen])
+
+
 def reduce(arr, factor=1, how=np.mean):
     """Reduces an array's size by a given factor.